+++ src/main/resources/templates/edit-product.html
@@ -5,7 +5,7 @@
 </head>
 <body>
 <h1>Edit Product</h1>
-<form th:action="@{/api/products/edit/{id}(id=${id})}" th:object="${product}" method="post">
+<form th:action="@{/admin/products/edit/{id}(id=${id})}" th:object="${product}" method="post">
     <input type="hidden" name="_method" value="put"/>
     <div>
         <label>Name:</label>
